requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
Faker==25.2.0
flake8==7.0.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
iniconfig==2.0.0
mccabe==0.7.0
multidict==6.0.5
outcome==1.3.0.post0
packaging==24.0
pluggy==1.4.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.2.0
PySocks==1.7.1
pytest==8.1.1
pytest-aiohttp==1.0.5
pytest-asyncio==0.23.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
selenium==4.19.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==1.4.50
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.2.1
webdriver-manager==4.0.1
wsproto==1.2.0
yarl==1.9.4